[REFACTOR] chang code from aiController to aiService
